+++ src/de/otto/tesla/stateful/metering.clj
@@ -3,25 +3,44 @@
     [com.stuartsierra.component :as component]
     [clojure.tools.logging :as log]
     [de.otto.goo.goo :as goo]
-    [de.otto.tesla.reporter.console :as prom-console]
-    [de.otto.tesla.reporter.graphite :as prom-graphite]
-    [de.otto.tesla.reporter.prometheus :as prom-endpoint]
-    [de.otto.tesla.reporter.graphite-dropwizard :as metrics-graphite-dropwizard]
-    [metrics.core :as dw]))
+    [de.otto.tesla.middleware.auth :as auth]
+    [de.otto.tesla.stateful.handler :as handlers]
+    [compojure.core :as c]
+    [overtone.at-at :as at]
+    [de.otto.tesla.stateful.scheduler :as sched]))
+
+(defn write-to-console []
+  (log/info "Metrics Reporting:\n"  (goo/text-format)))
+
+(defn start-console-reporter [console-config scheduler]
+  (let [interval-in-ms (* 1000 (:interval-in-s console-config))]
+    (log/info "Starting metrics console reporter")
+    (at/every interval-in-ms write-to-console (sched/pool scheduler) :desc "Console-Reporter")))
+
+(defn metrics-response [_]
+  (fn [_request] {:status  200
+                  :headers {"Content-Type" "text/plain"}
+                  :body    (goo/text-format)}))
+
+(defn- path-filter [metrics-path handler]
+  (c/GET metrics-path request (handler request)))
+
+(defn register-metrics-endpoint [{metrics-path :metrics-path :as config} handler authenticate-type authenticate-fn]
+  (log/info "Register metrics prometheus endpoint")
+  (handlers/register-handler handler ((->> metrics-response
+                                           (goo/timing-middleware)
+                                           (auth/wrap-auth authenticate-type authenticate-fn config)
+                                           (partial path-filter metrics-path)))))
 
 (defn- start-reporter! [handler scheduler authenticate-type prometheus-auth-fn [reporter-type reporter-config]]
   (case reporter-type
-    :console (prom-console/start! reporter-config scheduler)
-    :graphite (prom-graphite/start! reporter-config scheduler)
-    :prometheus (prom-endpoint/register-endpoint! reporter-config handler authenticate-type prometheus-auth-fn)
-    :graphite-dropwizard (metrics-graphite-dropwizard/start! dw/default-registry reporter-config)))
-
+    :console (start-console-reporter reporter-config scheduler)
+    :prometheus (register-metrics-endpoint reporter-config handler authenticate-type prometheus-auth-fn)))
 
 (defn- start-reporters! [config handler scheduler authenticate-type prometheus-auth-fn]
   (let [available-reporters (get-in config [:config :metrics])]
     (run! (partial start-reporter! handler scheduler authenticate-type prometheus-auth-fn) available-reporters)))
 
-;; Initialises a metrics-registry and a graphite reporter.
 (defrecord Metering [authenticate-type prometheus-auth-fn config handler scheduler]
   component/Lifecycle
   (start [self]
@@ -31,8 +50,7 @@
 
   (stop [self]
     (log/info "<- stopping metering")
-    ; TODO How to stop metrics graphite)
-    self))
+     self))
 
 (defn new-metering
   ([]