+++ CHANGES.md
@@ -8,6 +8,7 @@ This release cleans up remnants of past eperiments and unused functionality. Thi
 - Remove ```register-timed-handler``` from ```handler``` namespace. Use ```goo/timing-middleware``` instead.
 - Remove ```register-response-fn``` from handler namespace. This was only used internally.
 - Move internally used middlewares to separate namespaces.
+- Remove support for reporting via graphite. Only support prometheus reporting via goo and iapetos for the moment.
 
 
 ### 0.11.0